// src/app/shared/components/status/status.interface.ts
export enum StatusType {
  New = 'New',
  PendingReview = 'Pending Review',
  Completed = 'Completed',
  Active = 'Active',
  UnderReview = 'Under review',
  Inactive = 'Inactive',
}

export enum StatusClass {
  New = 'new',
  Pending = 'pending',
  Completed = 'completed',
  Inactive = 'inactive',
}